update to 1.1.4; remove redundant double quoting around filenames; fix erroneous triple quote in single line

+++ runRufus.sh
@@ -464,13 +464,13 @@ clean_up_files ()
 
     # Remove files from sub directories for this region only
 	if [ "$formatted_region" != "" ]; then
-		find ./Intermediates -maxdepth 1 -type f -name "*${formatted_region}*" -delete
-		find ./TempOverlap -maxdepth 1 -type f -name "*${formatted_region}*" -delete
-		find ./rufus_temp -maxdepth 1 -type f -name "*${formatted_region}*" -delete
 		find . -maxdepth 1 -type f -name "*${formatted_region}*generator*" -delete
 		find . -maxdepth 1 -type f -name "*${formatted_region}.txt" -delete
 		find . -maxdepth 1 -type p -name "*${formatted_region}*" -delete # Clean up pipes too
 		find . -maxdepth 1 -type f -name "*${formatted_region}.clean_vcf*" -delete
+		[ -d ./rufus_temp ] && find ./rufus_temp -maxdepth 1 -type f -name "*${formatted_region}*" -delete
+		[ -d ./TempOverlap ] && find ./TempOverlap -maxdepth 1 -type f -name "*${formatted_region}*" -delete
+		[ -d ./Intermediates ] && find ./Intermediates -maxdepth 1 -type f -name "*${formatted_region}*" -delete
 	fi
   else
     echo "not cleaning up files"
@@ -720,7 +720,7 @@ then
  # check for index file (needed for mpileup in post processing)
     if [[ ! -e "$_arg_subject".bai ]]
     then
-        echo "Index file for subject bam file "$_arg_subject" not found. Please place in data directory and rerun."
+        echo "Index file for subject bam file $_arg_subject not found. Please place in data directory and rerun."
         exit 1
     fi
     ProbandGenerator="${ProbandFileName}${region_postfix}.generator"
@@ -730,7 +730,7 @@ then
  # check for index file (needed for mpileup in post processing)
     if [[ ! -e "$_arg_subject".crai ]]
     then
-        echo "Index file for subject cram file "$_arg_subject" not found. Please place in data directory and rerun."
+        echo "Index file for subject cram file $_arg_subject not found. Please place in data directory and rerun."
         exit 1
     fi
 	if [ "$_arg_cramref" == "" ]
@@ -771,7 +771,7 @@ do
 		# check for index file (needed for mpileup in post processing)
 		if [[ ! -e "$parentFileName".bai ]]
 		then
-			echo "Index file for control bam file "$parentFileName" not found. Please place in data directory and rerun."
+			echo "Index file for control bam file $parentFileName not found. Please place in data directory and rerun."
 			exit 1
 		fi
 	    	parentGenerator="${parentFileName}${region_postfix}.generator"
@@ -782,7 +782,7 @@ do
 		# check for index file (needed for mpileup in post processing)
 		if [[ ! -e "$parentFileName".crai ]]
 		then
-			echo "Index file for control cram file "$parentFileName" not found. Please place in data directory and rerun."
+			echo "Index file for control cram file $parentFileName not found. Please place in data directory and rerun."
 			exit 1
 		fi
 		parentGenerator="${parentFileName}${region_postfix}.generator"
@@ -1186,7 +1186,7 @@ then
 	if [ $(head "$ProbandGenerator".Mutations.Mate1.fastq | wc -l | awk '{print $1}') -eq "0" ]; then
 		if [ -z $_arg_region ]
 		then
-			echo "ERROR: "No reads passed the filtering step in the entire genome. This is extremely unlikely and input files should be examined."
+			echo "ERROR: No reads passed the filtering step in the entire genome. This is extremely unlikely and input files should be examined."
 			exit 100
 		else
 			echo "No reads passed the filtering step in region $_arg_region. Stopping run."
@@ -1301,7 +1301,7 @@ if [ $( samtools view "${ProbandGenerator}".Mutations.fastq.bam | head | wc -l |
 			echo "ERROR: All reads failed to align to the reference genome. This is extremely unlikely for a whole genome run and something likely went wrong."
 			exit 100
 		else
-       		echo "WARNING: No reads aligned to the reference for "$ProbandGenerator".Mutations.fastq for the region $_arg_region. Stopping RUFUS run."
+       		echo "WARNING: No reads aligned to the reference for ${ProbandGenerator}.Mutations.fastq for the region $_arg_region. Stopping RUFUS run."
 			exit 0
 		fi
 fi 
@@ -1323,7 +1323,7 @@ else
 		_arg_refhash="empty"
 	fi
 
-    echo " bash  $RUFUSOverlap "$_arg_ref" "$ProbandGenerator".Mutations.fastq 5 $ProbandGenerator "$ProbandGenerator".k"$K"_c"$MutantMinCov".HashList "$K" "$Threads" "$_MaxAlleleSize" "$_arg_ref_bwa" "$rufus_invoc_file" "$_arg_refhash" "$ProbandGenerator".Jhash "$parentsString""
+    echo " bash  $RUFUSOverlap $_arg_ref ${ProbandGenerator}.Mutations.fastq 5 $ProbandGenerator ${ProbandGenerator}.k${K}_c${MutantMinCov}.HashList $K $Threads $_MaxAlleleSize $_arg_ref_bwa $rufus_invoc_file $_arg_refhash ${ProbandGenerator}.Jhash $parentsString"
     bash  $RUFUSOverlap "$_arg_ref" "$ProbandGenerator".Mutations.fastq 5 $ProbandGenerator "$ProbandGenerator".k"$K"_c"$MutantMinCov".HashList "$K" "$Threads" "$_MaxAlleleSize" "$_assemblySpeed" "$_arg_ref_bwa" "$rufus_invoc_file" "$_arg_refhash" "$ProbandGenerator".Jhash "$parentsString" 
     #bash  $RUFUSOverlap "$_arg_ref" "$ProbandGenerator".Mutations.fastq 3 $ProbandGenerator "$ProbandGenerator".k"$K"_c"$MutantMinCov".HashList "$K" "$Threads" "$_MaxAlleleSize" "$_assemblySpeed" "$ProbandGenerator".Jhash "$parentsString" "$_arg_ref_bwa" "$_arg_refhash"
     echo "Done with RUFUS overlap"
